requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
auth-token==0.0.3
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Django==4.2.13
django-otp==1.5.4
django-widget-tweaks==1.5.0
exceptiongroup==1.2.1
gTTS==2.5.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.35.10
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.5.2
PyJWT==2.9.0
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1

djangorestframework~=3.15.2